requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
twilio==8.10.0
requests==2.31.0
google-cloud-dialogflow==2.21.0
